fix(maven): use System.out redirection for SLF4J logging in Maven 3
Maven 3's slf4j-simple writes to System.out by default. Instead of
bypassing SLF4J with direct PrintStream writes, redirect System.out
to our capture stream during execution. This allows ExecutionEventLogger
to work natively with proper Maven-formatted output.

- Add System.out/err redirection in CachingMaven3Invoker.invoke()
- Create BatchExecutionListener extending ExecutionEventLogger
- Remove outputStream property from NxMaven3
- Delete BatchExecutionListenerPlexus (no longer needed)
